refactor(swe): collapse SWE evals to one real-harness example (#238)
There were three SWE eval examples: two oracle evals (swe.fix + swe.build)
that fed the grader the answer key to prove the *scorer* discriminates, and a
real-agent eval that drives a live LLM to solve blind. The oracle evals
measured the environment, not capability — and every discrimination assertion
they carried already lives in the pytest suite, exercised directly at the
grading seam (tests/test_swe_pack.py and tests/test_swe_build.py drive
check_success / check_feasibility / admit / episode_reward over gold +
known-bad variants, including the 1.00/0.60/0.00 reward triple).

So keep the one example that measures something unknown: rename
swe_agent_eval.py -> swe_eval.py as *the* SWE example — a real agent solves a
world blind, the held-out suite scores it, the harness/endpoint is the swap-in
seam. Delete both oracle evals and drop the now-moot TestEvalParity drift-guard
(it imported the deleted swe_build_eval); the units-only overlay it pinned still
lives in test_swe_build.py and is exercised by TestGrading/TestRewardSeam.
